requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
python-dotenv==1.0.0
pytz==2023.2
yarl==1.8.2
